# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ports",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/ports/read": {
            "get": {
                "description": "Retrieve a value from the specified input port by number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Input Ports"
                ],
                "summary": "Read value from an input port",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Port number",
                        "name": "number",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid port number",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ports/write": {
            "post": {
                "description": "Sends a transaction ID to the specified output port by port number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Output Ports"
                ],
                "summary": "Send a transaction ID to an output port",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Port number",
                        "name": "port_number",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "transaction_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid port number or transaction ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}